app/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
